Revert "forgot to change this"
This reverts commit bbc4d54f5b5a7f16c54253c100fa57b2650f0fbb.

+++ src/main/java/com/rabbitMq/rabbitmqscheduler/DTO/UserTransferOptions.java
@@ -17,6 +17,6 @@ public class UserTransferOptions {
     private Integer concurrencyThreadCount;
     private Integer parallelThreadCount;
     private Integer pipeSize;
-    private Long chunkSize;
+    private Integer chunkSize;
 
 }
